# sdk/openapi.json
{"openapi":"3.1.0","info":{"title":"Server","version":"1.0"},"paths":{"/agents/":{"get":{"tags":["agents"],"summary":"Get All Agents","description":"Get all agents (alive and dead) with metadata.","operationId":"get_all_agents_agents__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"$ref":"#/components/schemas/AgentInfo"},"type":"object","title":"Response Get All Agents Agents  Get"}}}}}}},"/agents/alive":{"get":{"tags":["agents"],"summary":"Get Alive Agents","description":"Get only currently alive agents.","operationId":"get_alive_agents_agents_alive_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"$ref":"#/components/schemas/AgentInfo"},"type":"object","title":"Response Get Alive Agents Agents Alive Get"}}}}}}},"/agents/dead":{"get":{"tags":["agents"],"summary":"Get Dead Agents","description":"Get only agents considered dead (missed heartbeat).","operationId":"get_dead_agents_agents_dead_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"$ref":"#/components/schemas/AgentInfo"},"type":"object","title":"Response Get Dead Agents Agents Dead Get"}}}}}}},"/agents/{agent_id}":{"get":{"tags":["agents"],"summary":"Get Agent","description":"Get detailed info about a specific agent.","operationId":"get_agent_agents__agent_id__get","parameters":[{"name":"agent_id","in":"path","required":true,"schema":{"type":"string","title":"Agent Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agents/{agent_id}/{module_name}":{"post":{"tags":["agents"],"summary":"Run Module","operationId":"run_module_agents__agent_id___module_name__post","parameters":[{"name":"agent_id","in":"path","required":true,"schema":{"type":"string","title":"Agent Id"}},{"name":"module_name","in":"path","required":true,"schema":{"type":"string","title":"Module Name"}},{"name":"untracked","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Untracked"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Module Request"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pipeline/":{"get":{"tags":["pipeline"],"summary":"Root","operationId":"root_pipeline__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/pipeline/runs/status/{measurement_id}":{"get":{"tags":["pipeline"],"summary":"Get Run Status","description":"Get the status of a specific run by its ID.","operationId":"get_run_status_pipeline_runs_status__measurement_id__get","parameters":[{"name":"measurement_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Measurement Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AgentInfo":{"properties":{"agent_id":{"type":"string","title":"Agent Id"},"alive":{"type":"boolean","title":"Alive"},"hostname":{"type":"string","title":"Hostname"},"last_seen":{"type":"string","format":"date-time","title":"Last Seen"},"config":{"additionalProperties":true,"type":"object","title":"Config"},"first_seen":{"type":"string","format":"date-time","title":"First Seen"},"total_heartbeats":{"type":"integer","title":"Total Heartbeats"}},"type":"object","required":["agent_id","alive","hostname","last_seen","first_seen","total_heartbeats"],"title":"AgentInfo"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}